test: move mockOauth2 client into hono folder
