+++ test/integration/helpers/test-client.ts
@@ -13,7 +13,7 @@ import {
 } from "../../../src/types";
 import { EmailAdapter } from "../../../src/adapters/interfaces/Email";
 import type { Email } from "../../../src/types/Email";
-import { mockOAuth2ClientFactory } from "../../../integration-test/mockOauth2Client";
+import { mockOAuth2ClientFactory } from "../mockOauth2Client";
 
 export async function getEnv() {
   const dialect = new SqliteDialect({
